requirements.txt
torch
pillow
transformers
fastapi
uvicorn[standard]
python-multipart
Jinja2